adding grpc server, configs, wire for DI

+++ internal/entity/interface.go
@@ -2,6 +2,6 @@ package entity
 
 type OrderRepositoryInterface interface {
 	Save(order *Order) error
-	GetTotal() (int, error)
+	GetTotal() (int64, error)
 	ListOrders() ([]Order, error)
 }
